Move archive dataSource bean out of archiveConfiguration
It caused circular dependency between datasource - metrics -
archiveConfiguration
